eval_service/requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
gunicorn==21.2.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
language_tool_python==3.2.2
mangum==0.20.0
MarkupSafe==3.0.3
nltk==3.9.2
numpy==2.4.0
packaging==25.0
psutil==7.2.1
pyspellchecker
pyphen==0.17.2
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers
setuptools==80.9.0
textstat==0.7.12
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.3
vaderSentiment==3.3.2
Werkzeug==3.1.5
